refactor: for locos without an explicit suppression detent, add 'or' logic with brake pipe reduction
It isn't always a good idea to require a lever position for suppression, particularly for NJT equipment with lapped brakes. But we can't rely entirely on brake pipe pressure, either, because the simulator can take several seconds to simulate that reduction. Adding an 'or' relationship allows a skilled player to enter suppression quickly and hold the train at 93 psi for softer braking action.